+++ test/runtests.jl
@@ -85,8 +85,10 @@ include("test_variational.jl")
 
 include("test_handlers.jl")
 include("test_ncrossbdm.jl")
-include("test_curl_lagc0d1_lagc0d2.jl")
+#include("test_curl_lagc0d1_lagc0d2.jl")
 include("test_gridfunction.jl")
+
+include("test_itsolver.jl")
 
 include("test_hh_lsvie.jl")